requirements.txt
gradio
numpy
pandas
scikit-learn
Pillow
joblib
